requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
supabase>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
